Good morning
I have been using ExplorerPatcher for over two years and I know how it works well.
I have been working since the beginning or almost on the translation of the software into French and I did it via the settings.reg and settings10.reg files locally on my PCs to create my .dlls derived from the official .dlls.
I have just noted with joy that the software is finally available in a multilingual version and if I was taken by surprise for the Fr translation due to lack of personal availability during the period of implementation of this improvement I would like to make my contribution to this project.
As with any translation, it is sometimes more judicious not to do the literal word for word and to transcribe the spirit of the initial text in a more understood form.
I found the proposed text excellent and very close to the one I have been using for months but I do however have some variations or even corrections to propose.
Not being at the initiative of the initial translation it is quite obvious to me that you are free to validate all or only part of my proposal.
I have currently only proposed the translation of the file for Windows 11 knowing that I have a similar proposal for Windows 10, but I prefer to wait for the validation of Windows 11 before transferring to Windows 10 for the sake of consistency of the two versions. (I use ExplorerPatcher on a personal fleet of 6 PCs running on different versions of Windows and I regularly measure what is happening for the evolution of EP in each system)
Above all, do not hesitate to give me feedback on my proposal because I remain open and it is by confronting divergences that we can approach relevance.

Very sincerely
Dennis

PS:
as for the expression "Remember the last section used in this window" I propose this version "Open ExplorerPatcher on the tab of the last section used" which is certainly far from the literal translation but which has the advantage of being better understood by ordinary users who discover the software.